requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary
python-dotenv
razorpay
flask-mail
authlib
requests
gunicorn
pandas
openpyxl
